# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Online market IPi lari",
    "version": "1.0.0",
    "description": "Bu shaxsiy online market web siti uchun qilingan loyiha IPI lari"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
...
    "/api/user/update": {
      "put": {
        "tags": ["User"],
        "summary": "Oddiy foydalanuvchi uchun ma'umotlar o'zgartirsh",
        "description": "Foydalanuvchilar uchun ma'lumotlar o'zgartirish",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "user",
            "schema": {
              "type": "object",
              "default": {
                "newPassword": "newPassword",
                "oldPassword": "oldPassword",
                "username": "username"
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "A successful response",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "default": {
                    "user": {
                      "email": "email",
                      "username": "username"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user/delete": {
      "delete": {
        "tags": ["User"],
        "summary": "Oddiy foydalanuvchi uchun tizimdan o'chib ketish",
        "description": "Foydalanuvchilar uchun tizimdan o'chib ketish",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "A successful response",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "default": {
                    "user": {
                      "email": "email",
                      "username": "username"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user/all": {
      "get": {
        "tags": ["User"],
        "summary": "Hamma userlar ma'lumotini olish uchun ",
        "description": "Ushbu APIdan faqat  admin roliga ega foydalanuvchilar foydalanadi",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "default": [
                    {
                      "product": "Product object"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/user/{id}": {
      "get": {
        "tags": ["User"],
        "summary": "Bitta foydalanuvchi ma'lumotlari olish",
        "description": "Ushbu APIdan faqat  admin roliga ega foydalanuvchilar foydalanadi",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "default": [
                    {
                      "product": "Product object"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sellers/create/{id}": {
      "post": {
        "tags": ["Sellers"],
        "summary": "Bitta Seller yaratish uchun  API",
        "description": "Sotuvchi foydalnuvchi bo'lib ro'yhatdan o'tgan bo'lishi kerak",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "For token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "siz qo'shmoqchi bo'lgan user id si",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "default": "server sorovi natijasi"
                }
              }
            }
          }
        }
      }
    },
    "/api/sellers/all": {
      "get": {
        "tags": ["Sellers"],
        "summary": "Hamma sotuvchilar ma'lumotini olish uchun ",
        "description": "Ushbu APIdan faqat  admin roliga ega foydalanuvchilar foydalanadi",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "default": [
                    {
                      "product": "Product object"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/sellers/{id}": {
      "get": {
        "tags": ["Sellers"],
        "summary": "Bitta sotuvchi ma'lumotlari olish",
        "description": "Ushbu APIdan faqat  admin roliga ega foydalanuvchilar foydalanadi",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "default": [
                    {
                      "product": "Product object"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}